Split the provider module
