+++ aws/rust-runtime/aws-config/external-types.toml
@@ -4,9 +4,9 @@
 # to the exposed SDK crates happens.
 allowed_external_types = [
    "aws_credential_types::cache::CredentialsCache",
-   "aws_credential_types::provider::ProvideCredentials",
-   "aws_credential_types::provider::Result",
-   "aws_credential_types::provider::SharedCredentialsProvider",
+   "aws_credential_types::provider::credentials::ProvideCredentials",
+   "aws_credential_types::provider::credentials::Result",
+   "aws_credential_types::provider::credentials::SharedCredentialsProvider",
    "aws_sdk_sts::types::_policy_descriptor_type::PolicyDescriptorType",
    "aws_smithy_async::rt::sleep::AsyncSleep",
    "aws_smithy_client::bounds::SmithyConnector",
