+++ aws/rust-runtime/aws-config/src/meta/credentials/chain.rs
@@ -17,7 +17,7 @@ use tracing::Instrument;
 /// * If a provider returns valid [`Credentials`](aws_credential_types::Credentials) they will be returned immediately.
 ///   No other credential providers will be used.
 /// * Otherwise, if a provider returns
-///   [`CredentialsError::CredentialsNotLoaded`](aws_credential_types::provider::error::CredentialsError::CredentialsNotLoaded),
+///   [`CredentialsError::CredentialsNotLoaded`](aws_credential_types::provider::credentials::error::CredentialsError::CredentialsNotLoaded),
 ///   the next provider will be checked.
 /// * Finally, if a provider returns any other error condition, an error will be returned immediately.
 ///
