+++ aws/rust-runtime/aws-credential-types/src/provider/token.rs
@@ -1,0 +1,6 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+//! AWS SDK Access Tokens
