+++ aws/rust-runtime/aws-inlineable/external-types.toml
@@ -1,5 +1,5 @@
 allowed_external_types = [
-    "aws_credential_types::provider::ProvideCredentials",
+    "aws_credential_types::provider::credentials::ProvideCredentials",
     "aws_endpoint::*",
     "aws_http::*",
     "aws_sig_auth::*",
