+++ aws/rust-runtime/aws-inlineable/src/no_credentials.rs
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use aws_credential_types::provider::{error::CredentialsError, future, ProvideCredentials};
+use aws_credential_types::provider::credentials::{
+    error::CredentialsError, future, ProvideCredentials,
+};
 
 /// Stub credentials provider for use when no credentials provider is used.
 #[non_exhaustive]
