+++ aws/rust-runtime/aws-sig-auth/src/lib.rs
@@ -79,7 +79,7 @@
 //! ## Sign a request for APIGateway execute-api
 //!
 //! ```no_run
-//! use aws_credential_types::provider::ProvideCredentials;
+//! use aws_credential_types::provider::credentials::ProvideCredentials;
 //! use aws_sig_auth::signer::{OperationSigningConfig, RequestConfig, SigV4Signer};
 //! use aws_smithy_http::body::SdkBody;
 //! use aws_types::region::{Region, SigningRegion};
